Change TileIndexInfo uv coordinates from int to float (though this should probably be reverted)

+++ include/Tileo/TileAtlas.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 
+
 #include <Tileo/TileIndexInfo.hpp>
 #include <allegro5/allegro.h>
 #include <vector>
@@ -25,7 +26,7 @@ namespace Tileo
 
       void load(ALLEGRO_BITMAP *tileset, int tile_width, int tile_height, int spacing=0);
       void draw_tile_to_atlas(ALLEGRO_BITMAP *tile, int tile_num, ALLEGRO_COLOR color=al_map_rgba_f(1, 1, 1, 1));
-      bool get_tile_uv(int index_num, int *u1, int *v1, int *u2, int *v2);
+      bool get_tile_uv(int index_num, float *u1, float *v1, float *u2, float *v2);
       ALLEGRO_BITMAP *get_bitmap(int index_num);
    };
 }
